backend/requirements.txt
fastapi
uvicorn
pydantic
beautifulsoup4
requests
google-cloud-aiplatform
python-multipart
anthropic[vertex]
python-dotenv
